fix(codes): move code calculations to payload_type
